Uncommitted work
Been a while since I've worked on this, so I honestly can't even
remember what I was doing when I stopped.

The robots.txt is kind of important to keep in place until moderation
tools are finished, though, because there are a fair number of impolite
names in the records and I'd rather they not be visible to the wider
internet until they've been cleaned up.